feat: complete phase 9 of post feature (Seeder & Factory)

+++ database/seeders/PostSeeder.php
@@ -12,8 +12,10 @@ class PostSeeder extends Seeder
      */
     public function run(): void
     {
-        Post::query()->delete();
+        if (!app()->isProduction()) {
+            Post::query()->delete();
+        }
 
-        Post::factory()->count(12)->create();
+        Post::factory()->count(15)->create();
     }
 }
